+++ drizzlepac/haputils/catalog_utils.py
@@ -956,9 +956,9 @@ class HAPCatalogBase:
         data_table.meta["Aper1 (arcsec)"] = self.aper_radius_arcsec[0]
         data_table.meta["Aper2 (arcsec)"] = self.aper_radius_arcsec[1]
         if proc_type == "segment":
-             data_table.meta["Threshold (sigma)"] = self._nsigma
+             data_table.meta["Threshold (sigma)"] = self.final_nsigma
         else:
-             data_table.meta["Theshold (sigma)"] = self.param_dict['nsigma']
+             data_table.meta["Threshold (sigma)"] = self.param_dict['nsigma']
 
         data_table.meta["Exposure Start"] = self.image.keyword_dict["expo_start"]
         data_table.meta["Total Exposure Time"] = self.image.keyword_dict["texpo_time"]
@@ -1637,6 +1637,9 @@ class HAPSegmentCatalog(HAPCatalogBase):
         # is deemed to be of poor quality, make sure to add documentation to the output catalog.
         self.is_big_island = False
 
+        # Final sigma value used based upon Gaussian or RickerWavelet processing
+        self.final_nsigma = None
+
 
     def identify_sources(self, **pars):
         """Use photutils to find sources in image based on segmentation.
@@ -1670,10 +1673,10 @@ class HAPSegmentCatalog(HAPCatalogBase):
             log.info("Image: {}".format(self.imgname))
             log.info("Gaussian FWHM (Filter for source detection in arcseconds): {}".format(self.param_dict['TWEAK_FWHMPSF']))
             log.info("size_source_box (no. of connected pixels needed for a detection): {}".format(self._size_source_box))
-            log.info("nsigma (threshold = nsigma * background_rms): {}".format(self._nsigma))
+            log.info("Gaussian nsigma (threshold = nsigma * background_rms): {} (Round 1), {} (Round2 for Bkg2D)".format(self._nsigma, 2.0 * self._nsigma))
             log.info("nlevels (no. of multi-thresholding levels for deblending): {}".format(self._nlevels))
             log.info("contrast (frac. flux for peak to be separate object, 0=max. deblend, 1=no deblend): {}".format(self._contrast))
-            log.info("RickerWavelet nsigma (threshold = nsigma * background_rms): {}".format(self._rw2d_nsigma))
+            log.info("RickerWavelet nsigma (threshold = nsigma * background_rms): {} (Round 1), {} (Round 2 for Bkg2D)".format(self._rw2d_nsigma, 2.0 * self._rw2d_nsigma))
             log.info("RickerWavelet kernel X- and Y-dimension: {}".format(self._rw2d_size))
             log.info("Pixel limit on biggest source (criterion for  RickerWavelet kernel): {}".format(self._rw2d_biggest_pixels))
             log.info("Percentage limit on biggest source (criterion for  RickerWavelet kernel): {}".format(100.0 * self._rw2d_biggest_source))
@@ -1888,6 +1891,7 @@ class HAPSegmentCatalog(HAPCatalogBase):
                             self.kernel = rw2d_kernel
                             segm_img = copy.deepcopy(rw_segm_img)
                             del rw_segm_img
+                            self.final_nsigma = rw2d_sigma_for_threshold
                         # The field was found to be crowded, but the biggest source second criterion is deemed OK.
                         elif (rw_is_big_crowded and (ratio_cg2rw_bigsource > self._ratio_bigsource_limit)):
                             log.info("The Round 2 of segmentation images may still contain big sources/islands.\n"
@@ -1897,6 +1901,7 @@ class HAPSegmentCatalog(HAPCatalogBase):
                             self.kernel = rw2d_kernel
                             segm_img = copy.deepcopy(rw_segm_img)
                             del rw_segm_img
+                            self.final_nsigma = rw2d_sigma_for_threshold
                         # The segmentation image is problematic and the big island/source fraction limits are exceeded,
                         # so deblending could take days, and the results would not be viable in any case.
                         else:
@@ -1914,6 +1919,7 @@ class HAPSegmentCatalog(HAPCatalogBase):
                         self.kernel = g2d_kernel
                         segm_img = copy.deepcopy(g_segm_img)
                         del g_segm_img
+                        self.final_nsigma = sigma_for_threshold
 
                 # The first round RickerWavelet segmentation image is good, continue with the processing
                 elif not rw_is_big_crowded and rw_segm_img:
@@ -1921,6 +1927,7 @@ class HAPSegmentCatalog(HAPCatalogBase):
                     segm_img = copy.deepcopy(rw_segm_img)
                     del rw_segm_img
                     del g_segm_img
+                    self.final_nsigma = self._rw2d_nsigma
 
                 # No segments were detected in the total data product - no further processing done for this TDP,
                 # but processing of another TDP should proceed.
@@ -1933,12 +1940,16 @@ class HAPSegmentCatalog(HAPCatalogBase):
                 self.kernel = g2d_kernel
                 segm_img = copy.deepcopy(g_segm_img)
                 del g_segm_img
+                self.final_nsigma = self._nsigma
 
             # No segments were detected in the total data product - no further processing done for this TDP,
             # but processing of another TDP should proceed.
             elif not g_segm_img:
                 self._define_empty_table(g_segm_img)
                 return
+
+            # Report the final nsigma value used to compute the threshold above which sources are detected
+            log.info("*** Final nsigma used (threshold = nsigma * background_rms): {}".format(self.final_nsigma))
 
             # If appropriate, deblend the segmentation image. Otherwise, use the current segmentation image
             ncount += 1
@@ -2000,13 +2011,14 @@ class HAPSegmentCatalog(HAPCatalogBase):
             log.info("{}".format("=" * 80))
             log.info("")
 
-        # This is the filter product section -  use sources identified in total detection product 
-        # previously generated
+        # This is the filter product section -  use sources and other information identified in
+        # the total detection product previously generated
         else:
             self.sources = self.tp_sources['segment']['sources']                   # segmentation image
-            self.kernel = self.tp_sources['segment']['kernel']                  
+            self.kernel = self.tp_sources['segment']['kernel']                     # image smoothing kernel
             self.total_source_table = self.tp_sources['segment']['source_cat']     # source catalog as a table
             self.total_source_cat = self.tp_sources['segment']['total_source_cat'] # SourceCatalog datatype
+            self.final_nsigma = self.tp_sources['segment']['final_nsigma']         # final sigma multiplicative factor used
 
         # For debugging purposes only, create a "regions" files to use for ds9 overlay of the segm_img.
         # Create the image regions file here in case there is a failure.  This diagnostic portion of the